Cleanup, fix tests
